+++ evm-contracts/src/v0.6/dev/AggregatorInterface.sol
@@ -10,8 +10,8 @@ interface AggregatorInterface is HistoricAggregatorInterface {
     returns (
       uint256 roundId,
       int256 answer,
-      uint64 startedAt,
-      uint64 updatedAt,
+      uint256 startedAt,
+      uint256 updatedAt,
       uint256 answeredInRound
     );
   function latestRoundData()
@@ -20,8 +20,8 @@ interface AggregatorInterface is HistoricAggregatorInterface {
     returns (
       uint256 roundId,
       int256 answer,
-      uint64 startedAt,
-      uint64 updatedAt,
+      uint256 startedAt,
+      uint256 updatedAt,
       uint256 answeredInRound
     );
 }
